backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic-settings==2.2.1
python-dotenv==1.0.1
supabase==2.4.3
playwright==1.44.0
beautifulsoup4==4.12.3
lxml==5.2.1
httpx==0.27.0
apscheduler==3.10.4
pytest==8.2.0
pytest-asyncio==0.23.6
